Move test to separate file, wasm can't support it yet
